Refactor Subjects so that it does not assume it shows subjects from one report.
