+++ components/frontend/src/subject/Subjects.js
@@ -1,6 +1,5 @@
-import React, { useContext } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
-import { Segment } from '../semantic_ui_react_wrappers';
 import {
     datePropType,
     datesPropType,
@@ -9,14 +8,8 @@ import {
     sortDirectionPropType,
     stringsPropType
 } from '../sharedPropTypes';
-import { DataModel } from '../context/DataModel';
-import { EDIT_REPORT_PERMISSION, ReadOnlyOrEditable } from '../context/Permissions';
-import { AddDropdownButton, CopyButton, MoveButton } from '../widgets/Button';
-import { add_subject, copy_subject, move_subject } from '../api/subject';
-import { subject_options } from '../widgets/menu_options';
 import { useDelayedRender } from '../utils';
 import { Subject } from './Subject';
-import { subjectTypes } from './SubjectType';
 
 export function Subjects({
     changed_fields,
@@ -28,8 +21,8 @@ export function Subjects({
     issueSettings,
     measurements,
     reload,
-    report,
     reports,
+    reportsToShow,
     report_date,
     sortColumn,
     sortDirection,
@@ -37,11 +30,16 @@ export function Subjects({
     visibleDetailsTabs
 }) {
     const visible = useDelayedRender();
-    const dataModel = useContext(DataModel)
-    const lastIndex = Object.keys(report.subjects).length - 1;
+    const subjects = []
+    reportsToShow.forEach((report) => {
+        Object.keys(report.subjects).forEach((subject_uuid) => {
+            subjects.push([report, subject_uuid])
+        })
+    })
+    const lastIndex = subjects.length - 1;
     return (
         <>
-            {Object.keys(report.subjects).map((subject_uuid, index) =>
+            {subjects.map(([report, subject_uuid], index) =>
                 visible || index < 3 ?
                     <Subject
                         changed_fields={changed_fields}
@@ -50,41 +48,22 @@ export function Subjects({
                         handleSort={handleSort}
                         hiddenColumns={hiddenColumns}
                         hiddenTags={hiddenTags}
-                        metricsToHide={metricsToHide}
                         issueSettings={issueSettings}
                         key={subject_uuid}
                         lastSubject={index === lastIndex}
                         measurements={measurements}
+                        metricsToHide={metricsToHide}
+                        reload={reload}
                         report={report}
-                        report_date={report_date}
                         reports={reports}
+                        report_date={report_date}
                         sortColumn={sortColumn}
                         sortDirection={sortDirection}
                         subject_uuid={subject_uuid}
                         toggleVisibleDetailsTab={toggleVisibleDetailsTab}
                         visibleDetailsTabs={visibleDetailsTabs}
-                        reload={reload}
                     /> : null
             )}
-            <ReadOnlyOrEditable requiredPermissions={[EDIT_REPORT_PERMISSION]} editableComponent={
-                <Segment basic>
-                    <AddDropdownButton
-                        item_type="subject"
-                        item_subtypes={subjectTypes(dataModel)}
-                        onClick={(subtype) => { add_subject(report.report_uuid, subtype, reload) }}
-                    />
-                    <CopyButton
-                        item_type="subject"
-                        onChange={(source_subject_uuid) => copy_subject(source_subject_uuid, report.report_uuid, reload)}
-                        get_options={() => subject_options(reports, dataModel)}
-                    />
-                    <MoveButton
-                        item_type="subject"
-                        onChange={(source_subject_uuid) => move_subject(source_subject_uuid, report.report_uuid, reload)}
-                        get_options={() => subject_options(reports, dataModel, report.report_uuid)}
-                    />
-                </Segment>}
-            />
         </>
     )
 }
@@ -98,8 +77,8 @@ Subjects.propTypes = {
     measurements: PropTypes.array,
     metricsToHide: metricsToHidePropType,
     reload: PropTypes.func,
-    report: PropTypes.object,
     reports: PropTypes.array,
+    reportsToShow: PropTypes.arrayOf(PropTypes.object),
     report_date: datePropType,
     sortColumn: PropTypes.string,
     sortDirection: sortDirectionPropType,
